"Fix" bridging for `*mut BOOL`.
This probably isn't the correct way to bridge `BOOL` pointers, but there's a problem: the `pBool` value wants an i16 ptr while `BOOL` carries an `i32`. It's almost certainly unsound to transmute pointers of different-sized types to make this work.

The current solution gets rid of the error on the Rust side, which is technically """sound""", according to how Rust is supposed to work. However, I have no clue if C++-side COM code actually pulls booleans from `i32` or `i16`s, so this could totally be wrong anyway. If they're wrong, we have to be wrong, too.